--- TrotWithTabs/app/src/main/java/com/example/trotwithtabs/GenreItem.java
@@ -1,25 +1,0 @@
-package com.example.trotwithtabs;
-
-public class GenreItem {
-
-    String name;
-
-    public GenreItem(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "GenreItem{" +
-                "name='" + name + '\'' +
-                '}';
-    }
-}
