+++ TrotWithTabs/app/src/main/java/com/trot/trotwithtabs/Genre.java
@@ -1,0 +1,228 @@
+package com.trot.trotwithtabs;
+
+import android.content.Context;
+import android.os.AsyncTask;
+import android.os.Bundle;
+import android.os.Parcelable;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.BaseAdapter;
+import android.widget.ListView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.google.api.client.googleapis.json.GoogleJsonResponseException;
+import com.google.api.client.http.HttpRequest;
+import com.google.api.client.http.HttpRequestInitializer;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.http.javanet.NetHttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.services.youtube.YouTube;
+import com.google.api.services.youtube.model.ResourceId;
+import com.google.api.services.youtube.model.SearchListResponse;
+import com.google.api.services.youtube.model.SearchResult;
+import com.google.api.services.youtube.model.Thumbnail;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
+public class Genre extends Fragment {
+    MainActivity activity;
+    Context context;
+
+    private static final String TAG = "genre";
+
+    private String result;
+    int genrePosition;
+    String[] list_genre = {"트로트 메들리","댄스트로트","전설의 트로트","최신 트로트","보이스트롯","미스트롯 송가인","미스터트롯","7080 메들리","7080 인기가수","강변가요제","대학가요제"};
+    ArrayList<GenreInfoList> genreInfoList;
+    ArrayList<GenreInfoList> genreInfoList2;
+
+    int querySize;
+
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        this.context = context;
+
+        activity = (MainActivity) getActivity();
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        querySize = 50;
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        ViewGroup rootView = (ViewGroup)inflater.inflate(R.layout.genre, container, false);
+
+        ListView listView = (ListView) rootView.findViewById(R.id.listViewGenre);
+
+        GenreAdapter adapter = new GenreAdapter();
+        for (int i = 0; i < list_genre.length; i++) {
+            adapter.addItem(new GenreItem(list_genre[i]));
+        }
+
+        listView.setAdapter(adapter);
+
+        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+
+
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view,
+                                    int position, long id) {
+                YoutubeAsyncTask youtubeAsyncTask = new YoutubeAsyncTask();
+                youtubeAsyncTask.execute();
+                genrePosition= ReturnGenrePosition(position);
+            }
+        });
+
+        return rootView;
+    }
+
+    public int ReturnGenrePosition(int position){
+        return position;
+    }
+
+    private class YoutubeAsyncTask extends AsyncTask<Void, Void, Void> {
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+        }
+
+        @Override
+        protected Void doInBackground(Void... voids) {
+            try {
+                HttpTransport HTTP_TRANSPORT = new NetHttpTransport();
+                final JsonFactory JSON_FACTORY = new JacksonFactory();
+                final long NUMBER_OF_VIDEOS_RETURNED = querySize;
+
+                YouTube youtube = new YouTube.Builder(HTTP_TRANSPORT, JSON_FACTORY, new HttpRequestInitializer() {
+                    public void initialize(HttpRequest request) throws IOException {
+                    }
+                }).setApplicationName("youtube-search-sample").build();
+
+                YouTube.Search.List search = youtube.search().list("id,snippet");
+
+                search.setKey(getString(R.string.youtube_api_key));
+
+                String genreName=list_genre[genrePosition];
+                //검색어 들어가는 부분
+                search.setQ(genreName);
+                // search.setChannelId("UCk9GmdlDTBfgGRb7vXeRMoQ"); //레드벨벳 공식 유투브 채널
+                search.setOrder("relevance"); //date relevance
+
+                search.setType("video");
+
+                search.setFields("items(id/kind,id/videoId,snippet/title,snippet/thumbnails/default/url)");
+                search.setMaxResults(NUMBER_OF_VIDEOS_RETURNED);
+                SearchListResponse searchResponse = search.execute();
+
+                List<SearchResult> searchResultList = searchResponse.getItems();
+
+                //검색어 들어가는 부분
+                if (searchResultList != null) {
+                    prettyPrint(searchResultList.iterator(), genreName);
+                }
+            } catch (GoogleJsonResponseException e) {
+                System.err.println("There was a service error: " + e.getDetails().getCode() + " : "
+                        + e.getDetails().getMessage());
+                System.err.println("There was a service error 2: " + e.getLocalizedMessage() + " , " + e.toString());
+            } catch (IOException e) {
+                System.err.println("There was an IO error: " + e.getCause() + " : " + e.getMessage());
+            } catch (Throwable t) {
+                t.printStackTrace();
+            }
+
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(Void aVoid) {
+            super.onPostExecute(aVoid);
+            Fragment genreDetailFragment = new GenreDetail();
+            genreInfoList=genreInfoList2;
+
+            Bundle bundle=new Bundle();
+            bundle.putParcelableArrayList("genreInfoList",(ArrayList<? extends Parcelable>) genreInfoList);
+            genreDetailFragment.setArguments(bundle);
+
+            ((MainActivity)getActivity()).getSupportFragmentManager().beginTransaction().replace(R.id.container, genreDetailFragment).addToBackStack(null).commit();
+        }
+
+        public void prettyPrint(Iterator<SearchResult> iteratorSearchResults, String query) {
+            if (!iteratorSearchResults.hasNext()) {
+                System.out.println(" There aren't any results for your query.");
+            }
+
+            StringBuilder sb = new StringBuilder();
+
+            genreInfoList2 = new ArrayList<>();
+
+            while (iteratorSearchResults.hasNext()) {
+                SearchResult singleVideo = iteratorSearchResults.next();
+                ResourceId rId = singleVideo.getId();
+
+                // Double checks the kind is video.
+                if (rId.getKind().equals("youtube#video")) {
+
+                    Thumbnail thumbnail = (Thumbnail) singleVideo.getSnippet().getThumbnails().get("default");
+                    genreInfoList2.add(new GenreInfoList(singleVideo.getSnippet().getTitle(),rId.getVideoId(),thumbnail.getUrl()));
+                    Log.d(TAG,genreInfoList2.get(0).title);
+                    Log.d(TAG,"하긴 함");
+                }
+
+            }
+
+            result = sb.toString();
+        }
+    }
+
+
+    class GenreAdapter extends BaseAdapter {
+        ArrayList<GenreItem> items = new ArrayList<GenreItem>();
+
+        @Override
+        public int getCount() {
+            return items.size();
+        }
+
+        public void addItem(GenreItem item){
+            items.add(item);
+            this.notifyDataSetChanged();
+        }
+
+        @Override
+        public Object getItem(int position) {
+            return items.get(position);
+        }
+
+        @Override
+        public long getItemId(int position) {
+            return position;
+        }
+
+        @Override
+        public View getView(int position, View convertView, ViewGroup parent) {
+            GenreItemView view = new GenreItemView(getContext());
+
+            GenreItem item = items.get(position);
+            view.setName(item.getName());
+            return view;
+        }
+
+        private void getApplicationContext() {
+        }
+    }
+
+}
