+++ TrotWithTabs/app/src/main/java/com/trot/trotwithtabs/MainActivity.java
@@ -1,0 +1,417 @@
+package com.trot.trotwithtabs;
+
+import androidx.annotation.RequiresApi;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+
+import android.content.Context;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.content.pm.Signature;
+import android.database.sqlite.SQLiteDatabase;
+import android.os.AsyncTask;
+import android.os.Build;
+import android.os.Bundle;
+import android.os.Parcelable;
+import android.util.Base64;
+import android.util.Log;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.BaseAdapter;
+import android.widget.SearchView;
+
+import com.google.android.material.tabs.TabLayout;
+import com.google.api.client.googleapis.json.GoogleJsonResponseException;
+import com.google.api.client.http.HttpRequest;
+import com.google.api.client.http.HttpRequestInitializer;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.http.javanet.NetHttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.services.youtube.YouTube;
+import com.google.api.services.youtube.model.ResourceId;
+import com.google.api.services.youtube.model.SearchListResponse;
+import com.google.api.services.youtube.model.SearchResult;
+import com.google.api.services.youtube.model.Thumbnail;
+
+import java.io.IOException;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
+public class MainActivity extends AppCompatActivity {
+    Home fragment0;
+    Genre fragment1;
+    Popular fragment2;
+    Singer fragment3;
+    Comedy fragment4;
+    Jjim fragment5;
+
+    Bundle mBundle;  //main bundle
+
+    DBOpenHelper helper;
+    SQLiteDatabase db;
+
+    MainActivity activity;
+    Context context;
+    private String result;
+
+    private static final String TAG = "singer";
+
+    MenuItem mSearch;
+    String searchStr;
+
+    ArrayList<SingerInfoList> singerInfoList;
+    ArrayList<SingerInfoList> singerInfoList2 = new ArrayList<>();
+
+    int querySize;
+
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        querySize=50;
+
+        getHashKey();
+        String keyHash = com.kakao.util.helper.Utility.getKeyHash(this /* context */);
+        Log.d(TAG,keyHash);
+
+        helper = new DBOpenHelper(this);
+        db = helper.getWritableDatabase();
+
+       Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayShowCustomEnabled(false);
+        getSupportActionBar().setTitle("");
+        getSupportActionBar().setIcon(R.drawable.logo_text);
+
+        fragment0 = new Home();
+        fragment1 = new Genre();
+        fragment3 = new Singer();
+        fragment4 = new Comedy();
+        fragment5 = new Jjim();
+
+        getSupportFragmentManager().beginTransaction().replace(R.id.container,fragment0).commit();
+
+        TabLayout tabs = (TabLayout) findViewById(R.id.tabs);
+        tabs.addTab(tabs.newTab().setText("홈"));
+        tabs.addTab(tabs.newTab().setText("장르"));
+        tabs.addTab(tabs.newTab().setText("인기"));
+        tabs.addTab(tabs.newTab().setText("가수"));
+        tabs.addTab(tabs.newTab().setText("예능"));
+        tabs.addTab(tabs.newTab().setText("찜"));
+
+        tabs.setOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
+            @Override
+            public void onTabSelected(TabLayout.Tab tab) {
+                int position =  tab.getPosition();
+
+                Fragment selected = null;
+                if(position == 0) {
+                    selected = fragment0;
+                    getSupportFragmentManager().beginTransaction().replace(R.id.container, selected).commit();
+                }else if(position==1){
+                    selected = fragment1;
+                    getSupportFragmentManager().beginTransaction().replace(R.id.container, selected).commit();
+                }else if(position==2){
+                    YoutubeAsyncTask youtubeAsyncTask = new YoutubeAsyncTask();
+                    youtubeAsyncTask.execute();
+                }else if(position==3){
+                    selected = fragment3;
+                    getSupportFragmentManager().beginTransaction().replace(R.id.container, selected).commit();
+                }else if(position==4){
+                    selected = fragment4;
+                    getSupportFragmentManager().beginTransaction().replace(R.id.container, selected).commit();
+                }else if(position==5){
+                    selected = fragment5;
+                    getSupportFragmentManager().beginTransaction().replace(R.id.container, selected).commit();
+                }
+            }
+            @Override
+            public void onTabUnselected(TabLayout.Tab tab) {
+            }
+
+            @Override
+            public void onTabReselected(TabLayout.Tab tab) {
+            }
+        });
+    }
+
+    public void clickHandler(View view) {
+
+    }
+    public boolean onCreateOptionsMenu(Menu menu){
+        super.onCreateOptionsMenu(menu);
+        getMenuInflater().inflate(R.menu.menu_main,menu);
+        mSearch = menu.findItem(R.id.menu_search);
+        mSearch.setOnActionExpandListener(new MenuItem.OnActionExpandListener() {
+            @Override
+            public boolean onMenuItemActionExpand(MenuItem item) {
+                return false;
+            }
+
+            @Override
+            public boolean onMenuItemActionCollapse(MenuItem item) {
+                return false;
+            }
+        });
+
+        SearchView sv = (SearchView)mSearch.getActionView();
+        sv.setSubmitButtonEnabled(true);
+        sv.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                Log.d("search",query);
+                searchStr=query;
+                YoutubeAsyncTask2 youtubeAsyncTask2 = new YoutubeAsyncTask2();
+                youtubeAsyncTask2.execute();
+                return true;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                return true;
+            }
+        });
+        return true;
+    }
+
+    public void fragBtnClick(Bundle bundle) {
+        this.mBundle = bundle;
+    } //fragBtnClcick()
+
+    private class YoutubeAsyncTask extends AsyncTask<Void, Void, Void> {
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+        }
+
+        @Override
+        protected Void doInBackground(Void... voids) {
+            try {
+                HttpTransport HTTP_TRANSPORT = new NetHttpTransport();
+                final JsonFactory JSON_FACTORY = new JacksonFactory();
+                final long NUMBER_OF_VIDEOS_RETURNED = querySize;
+
+                YouTube youtube = new YouTube.Builder(HTTP_TRANSPORT, JSON_FACTORY, new HttpRequestInitializer() {
+                    public void initialize(HttpRequest request) throws IOException {
+                    }
+                }).setApplicationName("youtube-search-sample").build();
+
+                YouTube.Search.List search = youtube.search().list("id,snippet");
+
+                search.setKey(getString(R.string.youtube_api_key));
+
+                search.setQ("트로트");
+                search.setOrder("relevance");
+
+                search.setType("video");
+
+                search.setFields("items(id/kind,id/videoId,snippet/title,snippet/thumbnails/default/url)");
+                search.setMaxResults(NUMBER_OF_VIDEOS_RETURNED);
+                SearchListResponse searchResponse = search.execute();
+
+                List<SearchResult> searchResultList = searchResponse.getItems();
+
+                if (searchResultList != null) {
+                    prettyPrint(searchResultList.iterator(), "트로트");
+                }
+            } catch (GoogleJsonResponseException e) {
+                System.err.println("There was a service error: " + e.getDetails().getCode() + " : "
+                        + e.getDetails().getMessage());
+                System.err.println("There was a service error 2: " + e.getLocalizedMessage() + " , " + e.toString());
+            } catch (IOException e) {
+                System.err.println("There was an IO error: " + e.getCause() + " : " + e.getMessage());
+            } catch (Throwable t) {
+                t.printStackTrace();
+            }
+
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(Void aVoid) {
+            super.onPostExecute(aVoid);
+            fragment2 = new Popular();
+            singerInfoList=singerInfoList2;
+
+            Bundle bundle=new Bundle();
+            bundle.putParcelableArrayList("singerInfoList",(ArrayList<? extends Parcelable>) singerInfoList);
+            fragment2.setArguments(bundle);
+
+            getSupportFragmentManager().beginTransaction().replace(R.id.container, fragment2).commit();
+        }
+
+        public void prettyPrint(Iterator<SearchResult> iteratorSearchResults, String query) {
+            if (!iteratorSearchResults.hasNext()) {
+                System.out.println(" There aren't any results for your query.");
+            }
+
+            StringBuilder sb = new StringBuilder();
+
+            while (iteratorSearchResults.hasNext()) {
+                SearchResult singleVideo = iteratorSearchResults.next();
+                ResourceId rId = singleVideo.getId();
+
+                if (rId.getKind().equals("youtube#video")) {
+
+                    Thumbnail thumbnail = (Thumbnail) singleVideo.getSnippet().getThumbnails().get("default");
+                    singerInfoList2.add(new SingerInfoList(singleVideo.getSnippet().getTitle(),rId.getVideoId(),thumbnail.getUrl()));
+                    Log.d(TAG,"하긴 함");
+                }
+            }
+            result = sb.toString();
+        }
+    }
+
+
+    class SingerAdapter extends BaseAdapter {
+        ArrayList<SingerItem> items = new ArrayList<SingerItem>();
+
+        @Override
+        public int getCount() {
+            return items.size();
+        }
+
+        public void addItem(SingerItem item){
+            items.add(item);
+        }
+
+        @Override
+        public Object getItem(int position) {
+            return items.get(position);
+        }
+
+        @Override
+        public long getItemId(int position) {
+            return position;
+        }
+
+        @Override
+        public View getView(int position, View convertView, ViewGroup parent) {
+            SingerItemView view = new SingerItemView(MainActivity.this.getApplicationContext());
+
+            SingerItem item = items.get(position);
+            view.setName(item.getName());
+            return view;
+        }
+
+        private void getApplicationContext() {
+        }
+    }
+    private void getHashKey(){
+        PackageInfo packageInfo = null;
+        try {
+            packageInfo = getPackageManager().getPackageInfo(getPackageName(), PackageManager.GET_SIGNATURES);
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+        }
+        if (packageInfo == null)
+            Log.e("KeyHash", "KeyHash:null");
+
+        for (Signature signature : packageInfo.signatures) {
+            try {
+                MessageDigest md = MessageDigest.getInstance("SHA");
+                md.update(signature.toByteArray());
+                Log.d("KeyHash", Base64.encodeToString(md.digest(), Base64.DEFAULT));
+            } catch (NoSuchAlgorithmException e) {
+                Log.e("KeyHash", "Unable to get MessageDigest. signature=" + signature, e);
+            }
+        }
+    }
+    private class YoutubeAsyncTask2 extends AsyncTask<Void, Void, Void> {
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+        }
+
+        @Override
+        protected Void doInBackground(Void... voids) {
+            try {
+                HttpTransport HTTP_TRANSPORT = new NetHttpTransport();
+                final JsonFactory JSON_FACTORY = new JacksonFactory();
+                final long NUMBER_OF_VIDEOS_RETURNED = querySize;
+
+                YouTube youtube = new YouTube.Builder(HTTP_TRANSPORT, JSON_FACTORY, new HttpRequestInitializer() {
+                    public void initialize(HttpRequest request) throws IOException {
+                    }
+                }).setApplicationName("youtube-search-sample").build();
+
+                YouTube.Search.List search = youtube.search().list("id,snippet");
+
+                search.setKey(getString(R.string.youtube_api_key));
+
+                search.setQ(searchStr);
+                search.setOrder("relevance");
+
+                search.setType("video");
+
+                search.setFields("items(id/kind,id/videoId,snippet/title,snippet/thumbnails/default/url)");
+                search.setMaxResults(NUMBER_OF_VIDEOS_RETURNED);
+                SearchListResponse searchResponse = search.execute();
+
+                List<SearchResult> searchResultList = searchResponse.getItems();
+
+                if (searchResultList != null) {
+                    prettyPrint2(searchResultList.iterator(), searchStr);
+                }
+            } catch (GoogleJsonResponseException e) {
+                System.err.println("There was a service error: " + e.getDetails().getCode() + " : "
+                        + e.getDetails().getMessage());
+                System.err.println("There was a service error 2: " + e.getLocalizedMessage() + " , " + e.toString());
+            } catch (IOException e) {
+                System.err.println("There was an IO error: " + e.getCause() + " : " + e.getMessage());
+            } catch (Throwable t) {
+                t.printStackTrace();
+            }
+
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(Void aVoid) {
+            super.onPostExecute(aVoid);
+            fragment2 = new Popular();
+            singerInfoList = singerInfoList2;
+
+            Bundle bundle=new Bundle();
+            bundle.putParcelableArrayList("singerInfoList",(ArrayList<? extends Parcelable>) singerInfoList);
+            fragment2.setArguments(bundle);
+
+            getSupportFragmentManager().beginTransaction().replace(R.id.container, fragment2).commit();
+        }
+
+        public void prettyPrint2(Iterator<SearchResult> iteratorSearchResults, String query) {
+            if (!iteratorSearchResults.hasNext()) {
+                System.out.println(" There aren't any results for your query.");
+            }
+
+            StringBuilder sb = new StringBuilder();
+
+            singerInfoList2 = new ArrayList<>();
+
+            while (iteratorSearchResults.hasNext()) {
+                SearchResult singleVideo = iteratorSearchResults.next();
+                ResourceId rId = singleVideo.getId();
+
+                // Double checks the kind is video.
+                if (rId.getKind().equals("youtube#video")) {
+
+                    Thumbnail thumbnail = (Thumbnail) singleVideo.getSnippet().getThumbnails().get("default");
+                    singerInfoList2.add(new SingerInfoList(singleVideo.getSnippet().getTitle(),rId.getVideoId(),thumbnail.getUrl()));
+                    Log.d(TAG,"하긴 함");
+                }
+            }
+            result = sb.toString();
+        }
+    }
+
+
+}
