+++ TrotWithTabs/app/src/main/java/com/trot/trotwithtabs/SingerItemView.java
@@ -1,0 +1,38 @@
+package com.trot.trotwithtabs;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.LayoutInflater;
+import android.widget.LinearLayout;
+import android.widget.TextView;
+
+import androidx.annotation.Nullable;
+
+public class SingerItemView extends LinearLayout {
+    TextView textView;
+
+    public SingerItemView(Context context) {
+        super(context);
+
+        init(context);
+    }
+
+    public SingerItemView(Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
+
+        init(context);
+    }
+
+
+
+    private void init(Context context){
+        LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        inflater.inflate(R.layout.singer_item, this, true);
+
+        textView = (TextView) findViewById(R.id.textView);
+    }
+
+    public void setName(String name){
+        textView.setText(name);
+    }
+}
