+++ TrotWithTabs/app/src/main/java/nodeal/example/you_tube_recyclerview/YoutubeGenre.java
@@ -26,15 +26,11 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.trotwithtabs.DBOpenHelper;
-import com.example.trotwithtabs.GenreDetail;
-import com.example.trotwithtabs.GenreDetailView;
-import com.example.trotwithtabs.GenreInfoList;
-import com.example.trotwithtabs.GenreItem;
-import com.example.trotwithtabs.R;
-import com.example.trotwithtabs.SingerDetail;
-import com.example.trotwithtabs.SingerInfoList;
-import com.example.trotwithtabs.SongJjimList;
+import com.trot.trotwithtabs.DBOpenHelper;
+import com.trot.trotwithtabs.GenreInfoList;
+import com.trot.trotwithtabs.GenreItem;
+import com.trot.trotwithtabs.R;
+import com.trot.trotwithtabs.SongJjimList;
 import com.kakao.kakaolink.v2.KakaoLinkResponse;
 import com.kakao.kakaolink.v2.KakaoLinkService;
 import com.kakao.message.template.ContentObject;
