+++ TrotWithTabs/app/src/main/java/nodeal/example/you_tube_recyclerview/YoutubeSinger.java
@@ -26,17 +26,13 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.trotwithtabs.DBOpenHelper;
-import com.example.trotwithtabs.R;
-import com.example.trotwithtabs.SingerDetail;
-import com.example.trotwithtabs.SingerDetailView;
-import com.example.trotwithtabs.SingerInfoList;
-import com.example.trotwithtabs.SingerItem;
-import com.example.trotwithtabs.SingerJjimList;
-import com.example.trotwithtabs.SongJjimList;
+import com.trot.trotwithtabs.DBOpenHelper;
+import com.trot.trotwithtabs.R;
+import com.trot.trotwithtabs.SingerInfoList;
+import com.trot.trotwithtabs.SingerItem;
+import com.trot.trotwithtabs.SongJjimList;
 import com.kakao.kakaolink.v2.KakaoLinkResponse;
 import com.kakao.kakaolink.v2.KakaoLinkService;
-import com.kakao.message.template.ButtonObject;
 import com.kakao.message.template.ContentObject;
 import com.kakao.message.template.FeedTemplate;
 import com.kakao.message.template.LinkObject;
