+++ TrotWithTabs/app/src/main/java/nodeal/example/you_tube_recyclerview/ui/YoutubeAdapter.java
@@ -1,15 +1,13 @@
 package nodeal.example.you_tube_recyclerview.ui;
 
-import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.trotwithtabs.R;
+import com.trot.trotwithtabs.R;
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.YouTubePlayer;
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.listeners.AbstractYouTubePlayerListener;
