+++ TrotWithTabs/app/src/main/java/nodeal/example/you_tube_recyclerview/ui/YoutubeViewHolder.java
@@ -1,13 +1,9 @@
 package nodeal.example.you_tube_recyclerview.ui;
 
-import android.content.Context;
-import android.os.Bundle;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
-import com.example.trotwithtabs.R;
+
+import com.trot.trotwithtabs.R;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.views.YouTubePlayerView;
